requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pyyaml>=6.0
pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.26.0
